Update Cache Script
