+++ workers/tiltify-cache/src/types/env.ts
@@ -9,6 +9,7 @@ export interface Env {
     FUNDRAISER_PUBLIC_ID: string;
     YOGSCAST_CAMPAIGN_SLUG: string;
     YOGSCAST_USERNAME_SLUG: string;
+    DURABLE_OBJECT_CACHE_KEY: string;
     REFRESH_TIME: number;
     ENABLE_REFRESH: boolean;
     ENABLE_GRAPH_REFRESH: boolean;
